requirements.txt
datasets
huggingface_hub
python-dotenv